package-lock and sidebar css working
