+++ components/CanvasContent.tsx
@@ -1,11 +1,12 @@
 import React from "react";
 import CanvasSizeBox from "./canvas/CanvasSizeBox";
-import CanvasSidebar from "./canvas/CanvasSidebar/CanvasSidebar"
+import CanvasSidebar from "./canvas/CanvasSidebar";
 
 class CanvasContent extends React.Component {
     render() {
         return <>
             <CanvasSizeBox />
+            <CanvasSidebar />
             </>;
     }
 }
