+++ components/canvas/CanvasSidebar.css
@@ -4,6 +4,7 @@
     top:0; 
     bottom:0;
     overflow-y: scroll;
+    background-color:red;
 
 }
 .content{
