+++ package.json
@@ -17,11 +17,13 @@
     "react-bootstrap": "1.0.0",
     "react-dom": "16.13.1",
     "react-redux": "7.2.0",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "styled-jsx": "^3.2.5"
   },
   "devDependencies": {
     "@types/node": "13.9.8",
     "@types/react": "16.9.29",
+    "file-loader": "^6.0.0",
     "typescript": "3.8.3"
   }
 }
